--- a/lib/reporting/line_items.rb
+++ b/lib/reporting/line_items.rb
@@ -17,7 +17,7 @@ module Reporting
 
     def list(line_item_includes = [variant: [:supplier, :product]])
       line_items = order_permissions.visible_line_items.in_orders(orders.result)
-        .order("supplier.name", "product.name", "variant.display_name")
+        .order("supplier.name", "product.name", "variant.display_name", "variant.unit_description")
 
       if @params[:supplier_id_in].present?
         line_items = line_items.supplied_by_any(@params[:supplier_id_in])