@drummer83 2025-02-18
Hi @mkllnk,
Thanks for this PR! I've tested the performance using Postman and the excellent explanation by @filipefurtad0 [here](https://ofn-user-guide.gitbook.io/ofn-testing-handbook/testers-toolkit/performance-testing-with-postman). 1000 Thanks! :heart: 

In the table below you can see the average run time for the Orders and Fulfillment reports.

| Report | Before | After |
| - | - | - |
| Order Cycle Supplier Totals (100 runs) | 483 ms | 516 ms |
| Order Cycle Supplier Totals by Distributor (100 runs) | 858 ms | 872 ms |
| Order Cycle Distributor Totals by Supplier (100 runs) | 1960 ms | 1965 ms |
| Order Cycle Customer Totals (10 runs) | 5223 ms | 5281 ms |

We can see a very slightly increased run time for all of them, but I think that's acceptable.

Thanks again and merging! :rocket: